// apps/providerservice/src/generated/prisma/models/ProviderService.ts
export type ProviderServiceWhereInput = {
  AND?: Prisma.ProviderServiceWhereInput | Prisma.ProviderServiceWhereInput[];
  OR?: Prisma.ProviderServiceWhereInput[];
  NOT?: Prisma.ProviderServiceWhereInput | Prisma.ProviderServiceWhereInput[];
  id?: Prisma.StringFilter<'ProviderService'> | string;
  providerId?: Prisma.StringFilter<'ProviderService'> | string;
  name?: Prisma.StringFilter<'ProviderService'> | string;
  description?: Prisma.StringNullableFilter<'ProviderService'> | string | null;
  category?: Prisma.StringNullableFilter<'ProviderService'> | string | null;
  price?:
    | Prisma.DecimalFilter<'ProviderService'>
    | runtime.Decimal
    | runtime.DecimalJsLike
    | number
    | string;
  durationMinutes?: Prisma.IntNullableFilter<'ProviderService'> | number | null;
  status?: Prisma.EnumProviderServiceStatusFilter<'ProviderService'> | $Enums.ProviderServiceStatus;
  createdAt?: Prisma.DateTimeFilter<'ProviderService'> | Date | string;
  updatedAt?: Prisma.DateTimeFilter<'ProviderService'> | Date | string;
  provider?: Prisma.XOR<Prisma.ProviderScalarRelationFilter, Prisma.ProviderWhereInput>;
  servicePersonProviderServices?: Prisma.ServicePersonProviderServiceListRelationFilter;
};